Close over the event name instead of function
For me on firefox, `f` would point to the last function in the dict, not the one
which came from name. With this change, we instead close over the name, and get
the function from dict at run time.

At least for me it now fires the correct event handler again.

The problem comes from calling `events` like this:

```
Template.sometemplate.events({
  'click .something': function,
  'dropped .something-esle': function,
  'click .menubar': function
});
```

So when there are more than one events in the `dict`.